solve day 7 both starts
-the first part was intuitive and took me a bit more than 20 min to solve.
however on the second part I missed a simple mistake in an my assamptions and it took me more than 2 hours to figure it out.
I assumed that || operator would only concat elements in the list of rules and not the current result of all of the previous operation. instead of left to right as it is stated in description I prioritazed || first. It is ironic becase I managed to implement solution to this problem and it is quite harder problem to solve than what is stated.